backend/requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication (Supabase JWT)
pyjwt[crypto]==2.8.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
httpx==0.25.1

# External APIs
requests==2.31.0
redis==5.0.1

# AWS S3 (optional)
boto3==1.29.7

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
